spring web
napravljen unos i ispis indeksa tjelesne mase

+++ PhysicalActivityMVC/src/main/java/com/josip/physical/activity/indeks_tjelesne_mase/BMIReprository.java
@@ -1,5 +1,5 @@
 package com.josip.physical.activity.indeks_tjelesne_mase;
 
 public interface BMIReprository {
-float izracunajBMI(float masa,float visina_osobe);
+double izracunajBMI(int masa,double visina_osobe);
 }
